moving tests from service-test.go to subnets_test.go and adding coverage for receiving bad existing node with higher seq
