+++ beacon-chain/p2p/subnets_test.go
@@ -10,14 +10,19 @@ import (
 	"github.com/OffchainLabs/prysm/v6/beacon-chain/cache"
 	"github.com/OffchainLabs/prysm/v6/beacon-chain/core/peerdas"
 	testDB "github.com/OffchainLabs/prysm/v6/beacon-chain/db/testing"
+	"github.com/OffchainLabs/prysm/v6/beacon-chain/p2p/peers"
+	"github.com/OffchainLabs/prysm/v6/beacon-chain/p2p/peers/scorers"
+	testp2p "github.com/OffchainLabs/prysm/v6/beacon-chain/p2p/testing"
 	"github.com/OffchainLabs/prysm/v6/cmd/beacon-chain/flags"
 	"github.com/OffchainLabs/prysm/v6/config/params"
 	ecdsaprysm "github.com/OffchainLabs/prysm/v6/crypto/ecdsa"
+	"github.com/OffchainLabs/prysm/v6/encoding/bytesutil"
 	"github.com/OffchainLabs/prysm/v6/testing/assert"
 	"github.com/OffchainLabs/prysm/v6/testing/require"
 	"github.com/ethereum/go-ethereum/p2p/enode"
 	"github.com/ethereum/go-ethereum/p2p/enr"
 	"github.com/libp2p/go-libp2p/core/crypto"
+	"github.com/libp2p/go-libp2p/core/network"
 	"github.com/prysmaticlabs/go-bitfield"
 )
 
@@ -540,4 +545,561 @@ func TestInitializePersistentSubnets(t *testing.T) {
 	assert.Equal(t, true, ok)
 	assert.Equal(t, 2, len(subs))
 	assert.Equal(t, true, expTime.After(time.Now()))
+}
+
+func TestFindPeersWithSubnets_NodeDeduplication(t *testing.T) {
+	params.SetupTestConfigCleanup(t)
+	cache.SubnetIDs.EmptyAllCaches()
+	defer cache.SubnetIDs.EmptyAllCaches()
+
+	ctx := context.Background()
+	db := testDB.SetupDB(t)
+
+	localNode1 := createTestNodeWithID(t, "node1")
+	localNode2 := createTestNodeWithID(t, "node2")
+	localNode3 := createTestNodeWithID(t, "node3")
+
+	// Create different sequence versions of node1 with subnet 1
+	setNodeSubnets(localNode1, []uint64{1})
+	setNodeSeq(localNode1, 1)
+	node1_seq1_subnet1 := localNode1.Node()
+	setNodeSeq(localNode1, 2)
+	node1_seq2_subnet1 := localNode1.Node() // Same ID, higher seq
+	setNodeSeq(localNode1, 3)
+	node1_seq3_subnet1 := localNode1.Node() // Same ID, even higher seq
+
+	// Node2 with different sequences and subnets
+	setNodeSubnets(localNode2, []uint64{1})
+	node2_seq1_subnet1 := localNode2.Node()
+	setNodeSubnets(localNode2, []uint64{2}) // Different subnet
+	setNodeSeq(localNode2, 2)
+	node2_seq2_subnet2 := localNode2.Node()
+
+	// Node3 with multiple subnets
+	setNodeSubnets(localNode3, []uint64{1, 2})
+	node3_seq1_subnet1_2 := localNode3.Node()
+
+	tests := []struct {
+		name             string
+		nodes            []*enode.Node
+		defectiveSubnets map[uint64]int
+		expectedCount    int
+		description      string
+		eval             func(t *testing.T, result []*enode.Node) // Custom validation function
+	}{
+		{
+			name: "No duplicates - unique nodes with same subnet",
+			nodes: []*enode.Node{
+				node2_seq1_subnet1,
+				node3_seq1_subnet1_2,
+			},
+			defectiveSubnets: map[uint64]int{1: 2},
+			expectedCount:    2,
+			description:      "Should return all unique nodes subscribed to subnet",
+			eval:             nil, // No special validation needed
+		},
+		{
+			name: "Duplicate with lower seq first - should replace",
+			nodes: []*enode.Node{
+				node1_seq1_subnet1,
+				node1_seq2_subnet1, // Higher seq, should replace
+				node2_seq1_subnet1, // Different node to ensure we process enough nodes
+			},
+			defectiveSubnets: map[uint64]int{1: 2}, // Need 2 peers for subnet 1
+			expectedCount:    2,
+			description:      "Should replace with higher seq node for same subnet",
+			eval: func(t *testing.T, result []*enode.Node) {
+				found := false
+				for _, node := range result {
+					if node.ID() == node1_seq2_subnet1.ID() && node.Seq() == node1_seq2_subnet1.Seq() {
+						found = true
+						break
+					}
+				}
+				require.Equal(t, true, found, "Should have node with higher seq")
+			},
+		},
+		{
+			name: "Duplicate with higher seq first - should keep existing",
+			nodes: []*enode.Node{
+				node1_seq3_subnet1, // Higher seq
+				node1_seq2_subnet1, // Lower seq, should be skipped (continue branch)
+				node1_seq1_subnet1, // Even lower seq, should also be skipped (continue branch)
+				node2_seq1_subnet1, // Different node
+			},
+			defectiveSubnets: map[uint64]int{1: 2},
+			expectedCount:    2,
+			description:      "Should keep existing node with higher seq and skip lower seq duplicates",
+			eval: func(t *testing.T, result []*enode.Node) {
+				found := false
+				for _, node := range result {
+					if node.ID() == node1_seq3_subnet1.ID() && node.Seq() == node1_seq3_subnet1.Seq() {
+						found = true
+						break
+					}
+				}
+				require.Equal(t, true, found, "Should have node with highest seq")
+			},
+		},
+		{
+			name: "Multiple updates for same node",
+			nodes: []*enode.Node{
+				node1_seq1_subnet1,
+				node1_seq2_subnet1, // Should replace seq1
+				node1_seq3_subnet1, // Should replace seq2
+				node2_seq1_subnet1, // Different node
+			},
+			defectiveSubnets: map[uint64]int{1: 2},
+			expectedCount:    2,
+			description:      "Should keep updating to highest seq",
+			eval: func(t *testing.T, result []*enode.Node) {
+				found := false
+				for _, node := range result {
+					if node.ID() == node1_seq3_subnet1.ID() && node.Seq() == node1_seq3_subnet1.Seq() {
+						found = true
+						break
+					}
+				}
+				require.Equal(t, true, found, "Should have node with highest seq")
+			},
+		},
+		{
+			name: "Duplicate with equal seq in subnets - should skip",
+			nodes: []*enode.Node{
+				node1_seq2_subnet1, // First occurrence
+				node1_seq2_subnet1, // Same exact node instance, should be skipped (continue branch)
+				node2_seq1_subnet1, // Different node
+			},
+			defectiveSubnets: map[uint64]int{1: 2},
+			expectedCount:    2,
+			description:      "Should skip duplicate with equal sequence number in subnet search",
+			eval: func(t *testing.T, result []*enode.Node) {
+				foundNode1 := false
+				foundNode2 := false
+				node1Count := 0
+				for _, node := range result {
+					if node.ID() == node1_seq2_subnet1.ID() {
+						require.Equal(t, node1_seq2_subnet1.Seq(), node.Seq(), "Node1 should have expected seq")
+						foundNode1 = true
+						node1Count++
+					}
+					if node.ID() == node2_seq1_subnet1.ID() {
+						foundNode2 = true
+					}
+				}
+				require.Equal(t, true, foundNode1, "Should have node1")
+				require.Equal(t, true, foundNode2, "Should have node2")
+				require.Equal(t, 1, node1Count, "Should have exactly one instance of node1")
+			},
+		},
+		{
+			name: "Mix with different subnets",
+			nodes: []*enode.Node{
+				node2_seq1_subnet1,
+				node2_seq2_subnet2, // Higher seq but different subnet
+				node3_seq1_subnet1_2,
+			},
+			defectiveSubnets: map[uint64]int{1: 2, 2: 1},
+			expectedCount:    2, // node2 (latest) and node3
+			description:      "Should handle nodes with different subnet subscriptions",
+			eval:             nil, // Basic count validation is sufficient
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			gFlags := new(flags.GlobalFlags)
+			gFlags.MinimumPeersPerSubnet = 1
+			flags.Init(gFlags)
+			defer flags.Init(new(flags.GlobalFlags))
+
+			fakePeer := testp2p.NewTestP2P(t)
+
+			s := &Service{
+				cfg: &Config{
+					MaxPeers: 30,
+					DB:       db,
+				},
+				genesisTime:           time.Now(),
+				genesisValidatorsRoot: bytesutil.PadTo([]byte{'A'}, 32),
+				peers: peers.NewStatus(ctx, &peers.StatusConfig{
+					PeerLimit:    30,
+					ScorerParams: &scorers.Config{},
+				}),
+				host: fakePeer.BHost,
+			}
+
+			localNode := createTestNodeRandom(t)
+
+			mockIter := testp2p.NewMockIterator(tt.nodes)
+			s.dv5Listener = testp2p.NewMockListener(localNode, mockIter)
+
+			digest, err := s.currentForkDigest()
+			require.NoError(t, err)
+
+			ctxWithTimeout, cancel := context.WithTimeout(ctx, 100*time.Millisecond)
+			defer cancel()
+
+			result, err := s.findPeersWithSubnets(
+				ctxWithTimeout,
+				AttestationSubnetTopicFormat,
+				digest,
+				1,
+				tt.defectiveSubnets,
+			)
+
+			require.NoError(t, err, tt.description)
+			require.Equal(t, tt.expectedCount, len(result), tt.description)
+
+			if tt.eval != nil {
+				tt.eval(t, result)
+			}
+		})
+	}
+}
+
+// TestFindPeersWithSubnets_FilterPeerRemoval tests that nodes failing filterPeer are removed when searching for subnet peers
+func TestFindPeersWithSubnets_FilterPeerRemoval(t *testing.T) {
+	// Setup test environment
+	params.SetupTestConfigCleanup(t)
+	cache.SubnetIDs.EmptyAllCaches()
+	defer cache.SubnetIDs.EmptyAllCaches()
+
+	ctx := context.Background()
+	db := testDB.SetupDB(t)
+
+	// Create LocalNodes
+	localNode1 := createTestNodeWithID(t, "node1")
+	localNode2 := createTestNodeWithID(t, "node2")
+	localNode3 := createTestNodeWithID(t, "node3")
+
+	// Create versions of node1 with subnet 1
+	setNodeSubnets(localNode1, []uint64{1})
+	setNodeSeq(localNode1, 1)
+	node1_seq1_valid_subnet1 := localNode1.Node()
+
+	// Create bad version (higher seq)
+	setNodeSeq(localNode1, 2)
+	node1_seq2_bad_subnet1 := localNode1.Node()
+
+	// Create another valid version
+	setNodeSeq(localNode1, 3)
+	node1_seq3_valid_subnet1 := localNode1.Node()
+
+	// Node2 with subnet 1
+	setNodeSubnets(localNode2, []uint64{1})
+	node2_seq1_valid_subnet1 := localNode2.Node()
+
+	// Node3 with subnet 1 and 2
+	setNodeSubnets(localNode3, []uint64{1, 2})
+	node3_seq1_valid_subnet1_2 := localNode3.Node()
+
+	tests := []struct {
+		name             string
+		nodes            []*enode.Node
+		defectiveSubnets map[uint64]int
+		expectedCount    int
+		description      string
+		eval             func(t *testing.T, result []*enode.Node)
+	}{
+		{
+			name: "Valid node in subnet followed by bad version - should remove",
+			nodes: []*enode.Node{
+				node1_seq1_valid_subnet1, // First add valid node with subnet 1
+				node1_seq2_bad_subnet1,   // Invalid version with higher seq - should delete
+				node2_seq1_valid_subnet1, // Different valid node with subnet 1
+			},
+			defectiveSubnets: map[uint64]int{1: 2}, // Need 2 peers for subnet 1
+			expectedCount:    1,                    // Only node2 should remain
+			description:      "Should remove node from map when bad version arrives, even if it has required subnet",
+			eval: func(t *testing.T, result []*enode.Node) {
+				foundNode1 := false
+				foundNode2 := false
+				for _, node := range result {
+					if node.ID() == node1_seq1_valid_subnet1.ID() {
+						foundNode1 = true
+					}
+					if node.ID() == node2_seq1_valid_subnet1.ID() {
+						foundNode2 = true
+					}
+				}
+				require.Equal(t, false, foundNode1, "Node1 should have been removed despite having subnet")
+				require.Equal(t, true, foundNode2, "Node2 should be present")
+			},
+		},
+		{
+			name: "Bad node with subnet stays bad even with higher seq",
+			nodes: []*enode.Node{
+				node1_seq2_bad_subnet1,   // First bad node - not added
+				node1_seq3_valid_subnet1, // Higher seq but same bad peer ID
+				node2_seq1_valid_subnet1, // Different valid node
+			},
+			defectiveSubnets: map[uint64]int{1: 2},
+			expectedCount:    1, // Only node2 (node1 remains bad)
+			description:      "Bad peer with subnet remains bad even with higher seq",
+			eval: func(t *testing.T, result []*enode.Node) {
+				foundNode1 := false
+				foundNode2 := false
+				for _, node := range result {
+					if node.ID() == node1_seq3_valid_subnet1.ID() {
+						foundNode1 = true
+					}
+					if node.ID() == node2_seq1_valid_subnet1.ID() {
+						foundNode2 = true
+					}
+				}
+				require.Equal(t, false, foundNode1, "Node1 should remain bad despite having subnet")
+				require.Equal(t, true, foundNode2, "Node2 should be present")
+			},
+		},
+		{
+			name: "Mixed valid and bad nodes with subnets",
+			nodes: []*enode.Node{
+				node1_seq1_valid_subnet1,   // Add valid node1 with subnet
+				node2_seq1_valid_subnet1,   // Add valid node2 with subnet
+				node1_seq2_bad_subnet1,     // Invalid update for node1 - should remove
+				node3_seq1_valid_subnet1_2, // Add valid node3 with multiple subnets
+			},
+			defectiveSubnets: map[uint64]int{1: 3}, // Need 3 peers for subnet 1
+			expectedCount:    2,                    // Only node2 and node3 should remain
+			description:      "Should handle removal of nodes with subnets when they become bad",
+			eval: func(t *testing.T, result []*enode.Node) {
+				foundNode1 := false
+				foundNode2 := false
+				foundNode3 := false
+				for _, node := range result {
+					if node.ID() == node1_seq1_valid_subnet1.ID() {
+						foundNode1 = true
+					}
+					if node.ID() == node2_seq1_valid_subnet1.ID() {
+						foundNode2 = true
+					}
+					if node.ID() == node3_seq1_valid_subnet1_2.ID() {
+						foundNode3 = true
+					}
+				}
+				require.Equal(t, false, foundNode1, "Node1 should have been removed")
+				require.Equal(t, true, foundNode2, "Node2 should be present")
+				require.Equal(t, true, foundNode3, "Node3 should be present")
+			},
+		},
+		{
+			name: "Node with subnet marked bad stays bad for all sequences",
+			nodes: []*enode.Node{
+				node1_seq1_valid_subnet1, // Add valid node1 with subnet
+				node1_seq2_bad_subnet1,   // Bad update - should remove and mark bad
+				node1_seq3_valid_subnet1, // Higher seq but still same bad peer ID
+				node2_seq1_valid_subnet1, // Different valid node
+			},
+			defectiveSubnets: map[uint64]int{1: 2},
+			expectedCount:    1, // Only node2 (node1 stays bad)
+			description:      "Once marked bad, subnet peer stays bad for all sequences",
+			eval: func(t *testing.T, result []*enode.Node) {
+				foundNode1 := false
+				foundNode2 := false
+				for _, node := range result {
+					if node.ID() == node1_seq3_valid_subnet1.ID() {
+						foundNode1 = true
+					}
+					if node.ID() == node2_seq1_valid_subnet1.ID() {
+						foundNode2 = true
+					}
+				}
+				require.Equal(t, false, foundNode1, "Node1 should stay bad")
+				require.Equal(t, true, foundNode2, "Node2 should be present")
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// Initialize flags for subnet operations
+			gFlags := new(flags.GlobalFlags)
+			gFlags.MinimumPeersPerSubnet = 1
+			flags.Init(gFlags)
+			defer flags.Init(new(flags.GlobalFlags))
+
+			// Create test P2P instance
+			fakePeer := testp2p.NewTestP2P(t)
+
+			// Create mock service
+			s := &Service{
+				cfg: &Config{
+					MaxPeers: 30,
+					DB:       db,
+				},
+				genesisTime:           time.Now(),
+				genesisValidatorsRoot: bytesutil.PadTo([]byte{'A'}, 32),
+				peers: peers.NewStatus(ctx, &peers.StatusConfig{
+					PeerLimit:    30,
+					ScorerParams: &scorers.Config{},
+				}),
+				host: fakePeer.BHost,
+			}
+
+			// Mark specific node versions as "bad" to simulate filterPeer failures
+			for _, node := range tt.nodes {
+				if node == node1_seq2_bad_subnet1 {
+					// Get peer ID from the node to mark it as bad
+					peerData, _, _ := convertToAddrInfo(node)
+					if peerData != nil {
+						s.peers.Add(node.Record(), peerData.ID, nil, network.DirUnknown)
+						// Mark as bad peer - this will make filterPeer return false
+						s.peers.Scorers().BadResponsesScorer().Increment(peerData.ID)
+						s.peers.Scorers().BadResponsesScorer().Increment(peerData.ID)
+						s.peers.Scorers().BadResponsesScorer().Increment(peerData.ID)
+					}
+				}
+			}
+
+			// Create local node for the listener
+			localNode := createTestNodeRandom(t)
+
+			// Create mock listener with iterator
+			mockIter := testp2p.NewMockIterator(tt.nodes)
+			s.dv5Listener = testp2p.NewMockListener(localNode, mockIter)
+
+			// Get fork digest for topic format
+			digest, err := s.currentForkDigest()
+			require.NoError(t, err)
+
+			// Run findPeersWithSubnets
+			ctxWithTimeout, cancel := context.WithTimeout(ctx, 100*time.Millisecond)
+			defer cancel()
+
+			result, err := s.findPeersWithSubnets(
+				ctxWithTimeout,
+				AttestationSubnetTopicFormat,
+				digest,
+				1,
+				tt.defectiveSubnets,
+			)
+
+			// Verify results
+			require.NoError(t, err, tt.description)
+			require.Equal(t, tt.expectedCount, len(result), tt.description)
+
+			// Run custom validation if provided
+			if tt.eval != nil {
+				tt.eval(t, result)
+			}
+		})
+	}
+}
+
+// callbackIterator allows us to execute callbacks at specific points during iteration
+type callbackIteratorForSubnets struct {
+	nodes     []*enode.Node
+	index     int
+	callbacks map[int]func() // map from index to callback function
+}
+
+func (c *callbackIteratorForSubnets) Next() bool {
+	// Execute callback before checking if we can continue (if one exists)
+	if callback, exists := c.callbacks[c.index]; exists {
+		callback()
+	}
+
+	return c.index < len(c.nodes)
+}
+
+func (c *callbackIteratorForSubnets) Node() *enode.Node {
+	if c.index >= len(c.nodes) {
+		return nil
+	}
+
+	node := c.nodes[c.index]
+	c.index++
+	return node
+}
+
+func (c *callbackIteratorForSubnets) Close() {
+	// Nothing to clean up for this simple implementation
+}
+
+func TestFindPeersWithSubnets_received_bad_existing_node(t *testing.T) {
+	// This test successfully triggers delete(nodeByNodeID, node.ID()) in subnets.go by:
+	// 1. Processing node1_seq1 first (passes filterPeer, gets added to map
+	// 2. Callback marks peer as bad before processing node1_seq2"
+	// 3. Processing node1_seq2 (fails filterPeer, triggers delete since ok=true
+	params.SetupTestConfigCleanup(t)
+	cache.SubnetIDs.EmptyAllCaches()
+	defer cache.SubnetIDs.EmptyAllCaches()
+
+	ctx := context.Background()
+	db := testDB.SetupDB(t)
+
+	// Create LocalNode with same ID but different sequences
+	localNode1 := createTestNodeWithID(t, "testnode")
+	setNodeSubnets(localNode1, []uint64{1})
+	node1_seq1 := localNode1.Node() // Get current node
+	currentSeq := node1_seq1.Seq()
+	setNodeSeq(localNode1, currentSeq+1) // Increment sequence by 1
+	node1_seq2 := localNode1.Node()      // This should have higher seq
+
+	// Additional node to ensure we have enough peers to process
+	localNode2 := createTestNodeWithID(t, "othernode")
+	setNodeSubnets(localNode2, []uint64{1})
+	node2 := localNode2.Node()
+
+	gFlags := new(flags.GlobalFlags)
+	gFlags.MinimumPeersPerSubnet = 1
+	flags.Init(gFlags)
+	defer flags.Init(new(flags.GlobalFlags))
+
+	fakePeer := testp2p.NewTestP2P(t)
+
+	service := &Service{
+		cfg: &Config{
+			MaxPeers: 30,
+			DB:       db,
+		},
+		genesisTime:           time.Now(),
+		genesisValidatorsRoot: bytesutil.PadTo([]byte{'A'}, 32),
+		peers: peers.NewStatus(ctx, &peers.StatusConfig{
+			PeerLimit:    30,
+			ScorerParams: &scorers.Config{},
+		}),
+		host: fakePeer.BHost,
+	}
+
+	// Create iterator with callback that marks peer as bad before processing node1_seq2
+	iter := &callbackIteratorForSubnets{
+		nodes: []*enode.Node{node1_seq1, node1_seq2, node2},
+		index: 0,
+		callbacks: map[int]func(){
+			1: func() { // Before processing node1_seq2 (index 1)
+				// Mark peer as bad before processing node1_seq2
+				peerData, _, _ := convertToAddrInfo(node1_seq2)
+				if peerData != nil {
+					service.peers.Add(node1_seq2.Record(), peerData.ID, nil, network.DirUnknown)
+					// Mark as bad peer - need enough increments to exceed threshold (6)
+					for i := 0; i < 10; i++ {
+						service.peers.Scorers().BadResponsesScorer().Increment(peerData.ID)
+					}
+				}
+			},
+		},
+	}
+
+	localNode := createTestNodeRandom(t)
+	service.dv5Listener = testp2p.NewMockListener(localNode, iter)
+
+	digest, err := service.currentForkDigest()
+	require.NoError(t, err)
+
+	// Run findPeersWithSubnets - node1_seq1 gets processed first, then callback marks peer bad, then node1_seq2 fails
+	ctxWithTimeout, cancel := context.WithTimeout(ctx, 1*time.Second)
+	defer cancel()
+
+	result, err := service.findPeersWithSubnets(
+		ctxWithTimeout,
+		AttestationSubnetTopicFormat,
+		digest,
+		1,
+		map[uint64]int{1: 2}, // Need 2 peers for subnet 1
+	)
+
+	require.NoError(t, err)
+	require.Equal(t, 1, len(result))
 }
